Create Data Upgrades main.yml
Data enhancement is all about making sure any data that is coming into the business is being looked at with a critical eye and is being filtered down to maximize its value. Imagine getting a set of numbers tossed at you without a clue as to what they mean and what they are going to do for you.